IRQ: x86: feat: Add kvm_set_irq function
Add a new function, kvm_set_irq, to the i8259.c file. The function is
responsible for setting the level of an IRQ line in the KVM virtual
machine.

The function uses the KVM_IRQ_LINE ioctl to set the IRQ level. If the
ioctl call fails, the program will exit with an error message.

This new function will be used to handle IRQ level changes in the i8259
PIC emulation code and inject IRQ to KVM.