PIT: x86: fix: Control timer in handle_pause
Move cpu_enable_ticks, cpu_disable_ticks and clock_enable call to
handle_pause to avoid repeated called.

In the handle_pause function, when resuming the VM, ticks and clock are
enabled if the irqchip mode is split. This ensures that the VM can continue
running with accurate timing.

Similarly, when pausing the VM, ticks and clock are disabled if the
irqchip mode is split. This prevents the VM from consuming unnecessary
CPU resources while paused and wrong time counting.